Tweaks in the Clustering methods  and in spatial pyramid in order for it to handle multiple dictionary models of different sizes
The Clustering method now keep count of its length, that is it knows how many clusters it holds.
The Spatial pyramid method was assuming that you could pass multiple dictionaries for it to use but all of them had the same size it now handles cases were the dictionaries are of different lengths.

Signed Off By <ricardo - rbk@dcc.ufmg.br>